requirements.txt
Flask
requests
requests_cache
cassandra-driver
